requirements.txt
openai>=0.27.0
pandas>=1.3.0
tenacity>=8.0.0
python-dotenv>=0.15.0